feat: Release redesign
As it turns out user and organization pages can only be served from the root
of the master branch so we had to tweak the setup a bit. Originally the plan was
to put all artifacts into the `/docs` directory but this also works. All source
files now live in `src` and is then build into the root directory.